Modularize Experience component and add Normalize css
I modified the Experience component to be better equipped to add further projects and jobs efficiently. Using props this component has been refactored to create additions with less tedium.

Normalize.css was added to give the page a standard feel on different browsers as well.

Lastly, I implemented AirBNB's style guide via eslint on the edited files. My goal is to do this to the entire project eventually.